+++ satellite/src/main/java/com/linbit/linstor/layer/storage/utils/DmStatCommands.java
@@ -77,7 +77,7 @@ public class DmStatCommands
                     "External command timed out" :
                     "External command threw an IOException",
                 null,
-                String.format("External command: %s", StringUtils.join(" ", command)),
+                String.format("External command: %s", StringUtils.joinShellQuote(command)),
                 exc
             );
         }