+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Added copy of GenDbTables to GenCrdV* classes for migrations
+- Escape executed commands in logs, making them paste-able
 
 ### Fixed